+++ allink_core/allink_base/models/models.py
@@ -22,7 +22,7 @@ from allink_core.allink_categories.models import AllinkCategory
 from allink_core.allink_base.models import Classes
 from allink_core.allink_base.models import AllinkBaseModelManager
 from allink_core.allink_base.models import AllinkMetaTagFieldsModel
-from allink_core.allink_base.models.choices import BLANK_CHOICE, TITLE_CHOICES, H1
+from allink_core.allink_base.models.choices import TITLE_CHOICES, H1
 
 
 @python_2_unicode_compatible
@@ -83,7 +83,7 @@ class AllinkBaseModel(AllinkMetaTagFieldsModel):
         AllinkCategory,
         blank=True
     )
-    active = models.BooleanField(
+    is_active = models.BooleanField(
         _(u'Active'),
         default=True
     )
@@ -105,7 +105,7 @@ class AllinkBaseModel(AllinkMetaTagFieldsModel):
 
     @classmethod
     def get_published(cls):
-        return cls.objects.filter(active=True)
+        return cls.objects.filter(is_active=True)
 
     @classmethod
     def get_next(cls, instance):
@@ -122,7 +122,7 @@ class AllinkBaseModel(AllinkMetaTagFieldsModel):
         """
         result = AllinkCategory.objects.none()
         for root in AllinkCategory.get_root_nodes().filter(model_names__contains=[cls._meta.model_name]):
-            result |= root.get_children()
+            result |= root.get_descendants()
         return result
 
     @property
@@ -594,7 +594,7 @@ class AllinkBaseAppContentPlugin(AllinkBasePlugin):
             filters = [((None, _(u'All'),))]
             fk_model, is_translated = self.get_field_info(fieldname)
             if fk_model:
-                filters.extend((entry.id, entry.__unicode__(),) for entry in
+                filters.extend((entry.id, entry.__str__(),) for entry in
                                fk_model.objects.filter(
                                    id__in=self.get_distinct_values_of_field(fieldname)))
             # field is no foreignkey and no m2m