+++ allink_core/allink_base/utils/__init__.py
@@ -1,2 +1,2 @@
-from allink_core.allink_base.utils.utils import base_url, get_additional_templates, get_additional_choices, get_height_from_ratio, get_project_color_choices, get_ratio_choices  # noqa
+from allink_core.allink_base.utils.utils import base_url, get_additional_templates, get_additional_choices, get_height_from_ratio, get_project_color_choices, get_ratio_choices, get_ratio_choices_orig # noqa
 from allink_core.allink_base.utils.cms_toolbars import AllinkBaseModifierMixin  # noqa